BYUT

+++ Backend/database/migrations/2024_05_02_133148_create_postes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('libelle');
             $table->integer('duree_kangurou')->nullable();
-            $table->decimal('montant_kangurou', 10, 2)->nullable();
+            $table->integer('montant_kangurou')->nullable();
             $table->timestamps();
         });
     }
